requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dlib==19.21.0
face-recognition-models==0.3.0
geojson==2.5.0
idna==2.10
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
pyowm==3.0.0
PySocks==1.7.1
pyTelegramBotAPI==3.7.3
requests==2.24.0
scipy==1.5.2
six==1.15.0
urllib3==1.25.10
